created houstonBrand module (client rendering) left rest of footer to render server side
